requirements.txt
flask
flask-limiter==3.5.0
python-dotenv==1.0.0
pinecone-client==3.0.1
huggingface_hub==0.19.4
pypdf==3.17.0
requests==2.31.0
flask-wtf==1.1.1
wtforms==3.0.1
langchain-core
PyPDF2
langchain-text-splitters
pinecone 
openai 
tiktoken 
langchain-community